GVR-SceneKit
This is an example of using Apple SceneKit with iOS Google Virtual
Reality SDK for using with Google Cardboard.

Currently (early June 2016 April 2017) it is the easiest way to develop
VR app on the iPhone. Hope this is going to change after WWDC 2016 2017.

The project codename is ’VRBoilerplate’ and initially it was prepared
for iOS programming competition for uadevchallenge 2016. It supports
both Objective-C and Swift.